refactor(graph): extract store schema foundations
